encoding: Move Redundancy from maccommands to types module
