+++ lorawan-encoding/src/maccommandcreator.rs
@@ -1,5 +1,5 @@
-use super::maccommands::{mac_commands_len, DataRateRange, Redundancy, SerializableMacCommand};
-use crate::types::{ChannelMask, DLSettings, Frequency};
+use super::maccommands::{mac_commands_len, DataRateRange, SerializableMacCommand};
+use crate::types::{ChannelMask, DLSettings, Frequency, Redundancy};
 
 #[derive(Debug, PartialEq)]
 #[cfg_attr(feature = "defmt-03", derive(defmt::Format))]
@@ -233,7 +233,7 @@ impl LinkADRReqCreator {
     ///
     /// # Argument
     ///
-    /// * redundancy - instance of maccommands::Redundancy or anything that can
+    /// * redundancy - instance of types::Redundancy or anything that can
     ///   be converted into it.
     pub fn set_redundancy<T: Into<Redundancy>>(&mut self, redundancy: T) -> &mut Self {
         let converted = redundancy.into();
