+++ lorawan-encoding/src/maccommands.rs
@@ -10,6 +10,9 @@ pub use crate::types::DLSettings;
 #[deprecated(note = "Use lorawan::types::Frequency")]
 pub use crate::types::Frequency;
 
+#[deprecated(note = "Use lorawan::types::Redundancy")]
+pub use crate::types::Redundancy;
+
 #[derive(Debug, PartialEq)]
 #[cfg_attr(feature = "defmt-03", derive(defmt::Format))]
 pub enum Error {
@@ -292,38 +295,6 @@ impl LinkADRReqPayload<'_> {
 impl<'a> From<&'a [u8; 4]> for LinkADRReqPayload<'a> {
     fn from(v: &'a [u8; 4]) -> Self {
         LinkADRReqPayload(v)
-    }
-}
-
-/// Redundancy represents the LinkADRReq Redundancy from LoRaWAN.
-#[derive(Debug, PartialEq, Eq)]
-pub struct Redundancy(u8);
-
-impl Redundancy {
-    /// Constructs a new Redundancy from the provided data.
-    pub fn new(data: u8) -> Self {
-        Redundancy(data)
-    }
-
-    /// Controls the interpretation of the previously defined ChannelMask bit mask.
-    pub fn channel_mask_control(&self) -> u8 {
-        (self.0 >> 4) & 0x07
-    }
-
-    /// How many times each message should be repeated.
-    pub fn number_of_transmissions(&self) -> u8 {
-        self.0 & 0x0f
-    }
-
-    /// The integer value of the Redundancy.
-    pub fn raw_value(&self) -> u8 {
-        self.0
-    }
-}
-
-impl From<u8> for Redundancy {
-    fn from(v: u8) -> Self {
-        Redundancy(v)
     }
 }
